test(inline-lane): bd-yom8c round-1 — reconcile 144 inline #[cfg(test)] modules; lift import-masking (visible 1319→922)
Reconciled all 144 enum1-visible inline-test files across every src area to the
current prod API via per-file parallel subagents (test-code only; preserve intent;
genuinely-removed-API modules gated #[cfg(any())] + // FIXME(bd-yom8c)). 101/144
files now compile clean; 43 retain residual cascades.

KEY FINDING: enum1's 1319 was an UNDERCOUNT. Failed `use super::{...}` imports
(E0432) suppress all downstream "cannot find in scope" errors (E0425/E0433/E0422)
in that module. Fixing the imports surfaces the hidden layer — e.g.
supply_chain/artifact_signing 1→148, sdk/verifier_sdk 0→66 (never edited; pure
unmask), several control_plane modules 0→N. enum3 shows 922 true errors / 69 files
with only E0432=2 remaining, so masking is essentially fully lifted.

Net visible 1319→922 (-30%) with full visibility for round 2. Lane remains
BINARY-red until all files compile. Round 2 targets the now-visible residual
(43 partially-fixed + 26 newly-surfaced/unmasked files).

+++ crates/franken-node/src/connector/lease_coordinator.rs
@@ -1197,7 +1197,7 @@ mod tests {
         let cands = vec![
             CoordinatorCandidate {
                 node_id: "node-a".to_string(),
-                weight: u32::MAX,
+                weight: u64::MAX,
             },
             CoordinatorCandidate {
                 node_id: "node-b".to_string(),
@@ -1265,7 +1265,7 @@ mod tests {
         let cands = vec![
             CoordinatorCandidate {
                 node_id: "node-a".to_string(),
-                weight: u32::MAX,
+                weight: u64::MAX,
             },
             CoordinatorCandidate {
                 node_id: "node-b".to_string(),
